backend/requirements.txt
# Phase 1: Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
google-play-scraper>=1.2.4
langdetect>=1.0.9

# Phase 2 & 3: LLM
groq>=0.5.0

# Phase 4: Email (Gmail MCP)
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0

# Utilities
markdown>=3.5
python-dotenv>=1.0.0
structlog>=24.1.0
apscheduler>=3.10.0
pyyaml>=6.0.1

# Testing
pytest>=8.0.0
pytest-mock>=3.12.0
